Module type M_tree.T

type 'a t

val create : ?branching_factor:int -> ?eq:( 'a -> 'a -> bool ) -> ( 'a -> 'a -> float ) -> 'a t

Create an M-tree.

parameter branching_factor
The branching factor of the tree (default: 32).

parameter eq
An optional equality function for tree elements. If none is provided, then the distance function will be used.

parameter distance
Distance metric between tree elements.

val iter : 'a t -> 'a Iter.t

Iterate over the contents of the tree.

val insert : 'a t -> 'a -> unit

Insert a new point into the tree.

val range : 'a t -> 'a -> radius:float -> 'a Iter.t

Find all points that are within some radius of a query point.